ts_to_py_codes/embedding/requirements.txt
openai==1.12.0
qdrant-client==1.7.0
python-dotenv==1.0.0 